updateorganization.php	-	Changed the html to display the values of existing entries in the database to be double quote delimited instead of single quote (this now displays apostrophes) updateorganization2.php	-	Added mysql_real_escape_string() to the posted variables to automatically escape the input characters.  This fixes the issue with apostrophes.
updateperson.php	-	Changed the html to display the values of existing entries in the database to be double quote delimited instead of single quote (this now displays apostrophes)
updateperson2.php	-	Added mysql_real_escape_string() to the posted variables to automatically escape the input characters.  This fixes the issue with apostrophes.